requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-mail==0.9.1
flask-wtf==1.2.1
python-dotenv==1.0.0
sendgrid==6.10.0
werkzeug==3.0.1
pytz==2023.3
